category_update

+++ ecommerce/product/views.py
@@ -538,9 +538,11 @@ class ProductCategoryUpdateView(APIView):
         try:
             product = request.data.get('product')
             categories_list = request.data.get('categories')
-            for item in CategoryProduct.objects.filter(product_id=product):
-                if item.category_id not in categories_list:
-                    item.delete()
+            category_products = CategoryProduct.objects.filter(product_id=product)
+            for item in category_products:
+                if category_products.count() > 1:
+                    if item.category_id not in categories_list:
+                        item.delete()
             for category in categories_list:
                 CategoryProduct.objects.get_or_create(category_id=category,
                                                      product_id=product)
